+++ lib/modules/network/network_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: network
-// Updated at: 2026-06-07 12:40:37
-// Commit: chore(network): update dependencies and cleanup unused imports in network
+// Updated at: 2026-06-07 16:53:53
+// Commit: style(network): harmonize light and dark theme palette for network
 
 class networkComponentState {
-  final String commitRef = '1833';
-  final String lastMessage = 'chore(network): update dependencies and cleanup unused imports in network';
-  final DateTime timestamp = DateTime.parse('2026-06-07T12:40:37+05:00');
+  final String commitRef = '1844';
+  final String lastMessage = 'style(network): harmonize light and dark theme palette for network';
+  final DateTime timestamp = DateTime.parse('2026-06-07T16:53:53+05:00');
 }